Refactor processing code for reuse
match specific to non-specific
